+++ SimulationSys/user.h
@@ -12,8 +12,8 @@ class User
 private:
 	int sno;
 	string id;
-	float lon;
-	float lat;
+	double lon;
+	double lat;
 	string isEnabled;
 	string isOnDuty;
 	string name;
@@ -23,8 +23,8 @@ private:
 public:
 	int get_sno() const { return sno; };
 	string get_id() const { return id; };
-	float get_lon() const { return lon; };
-	float get_lat() const { return lat; };
+	double get_lon() const { return lon; };
+	double get_lat() const { return lat; };
 	string get_isEnabled() const { return isEnabled; };
 	string get_isOnDuty() const { return isOnDuty; };
 	string get_name() const { return name; };
@@ -32,8 +32,8 @@ public:
 
 	void set_sno(int s) { sno = s; };
 	void set_id(string i) { id = i; };
-	void set_lon(float lo) { lon = lo; };
-	void set_lat(float la) { lat = la; };
+	void set_lon(double lo) { lon = lo; };
+	void set_lat(double la) { lat = la; };
 	void set_isEnabled(string e) { isEnabled = e; };
 	void set_isOnDuty(string o) { isOnDuty = o; };
 	void set_name(string n) { name = n; };
@@ -42,7 +42,7 @@ public:
 	void show();
 
 	User();
-	User(int s, const string & i, float lo, float la,
+	User(int s, const string & i, double lo, double la,
 		const string & e, const string & o, const string & n,
 		const string & m);
 	~User();
